+++ Packages/OpenApiCodeGen/Editor/Domain/Progress/FailedProgressStatus.cs
@@ -1,5 +1,5 @@
 ﻿#nullable enable
-namespace ReBeat.OpenApiCodeGen.UI
+namespace Rhycol.OpenApiCodeGen.UI
 {
     internal class FailedProgressStatus : IProgressStatus
     {