+++ Packages/OpenApiCodeGen/Editor/Domain/Progress/IProgressStatus.cs
@@ -1,5 +1,5 @@
 ﻿#nullable enable
-namespace ReBeat.OpenApiCodeGen.UI
+namespace Rhycol.OpenApiCodeGen.UI
 {
     internal interface IProgressStatus
     {