+++ Packages/OpenApiCodeGen/Editor/Service/Dto/CheckDockerInstalledDto.cs
@@ -1,4 +1,4 @@
-﻿namespace ReBeat.OpenApiCodeGen.Core
+﻿namespace Rhycol.OpenApiCodeGen.Core
 {
     internal class CheckDockerInstalledDto
     {