+++ Packages/OpenApiCodeGen/Editor/Service/Dto/GenerateApiClientDto.cs
@@ -1,7 +1,7 @@
 ﻿#nullable enable
-using ReBeat.OpenApiCodeGen.Core;
+using Rhycol.OpenApiCodeGen.Core;
 
-namespace ReBeat.OpenApiCodeGen.UI
+namespace Rhycol.OpenApiCodeGen.UI
 {
     internal class GenerateApiClientDto
     {