+++ Packages/OpenApiCodeGen/Editor/Service/Dto/SetupDto.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace ReBeat.OpenApiCodeGen.Core
+namespace Rhycol.OpenApiCodeGen.Core
 {
     internal class SetupDto
     {