+++ Packages/OpenApiCodeGen/Editor/Share/Lib/OpenApi/OpenApiDependenceLibrary.cs
@@ -1,6 +1,6 @@
-using ReBeat.OpenApiCodeGen.Core;
+using Rhycol.OpenApiCodeGen.Core;
 
-namespace ReBeat.OpenApiCodeGen.UI
+namespace Rhycol.OpenApiCodeGen.UI
 {
     public enum OpenApiDependenceLibrary
     {