+++ Packages/OpenApiCodeGen/Editor/Share/Lib/OpenApi/OpenApiTargetFramework.cs
@@ -1,7 +1,7 @@
 ﻿using System;
-using ReBeat.OpenApiCodeGen.Core;
+using Rhycol.OpenApiCodeGen.Core;
 
-namespace ReBeat.OpenApiCodeGen.UI
+namespace Rhycol.OpenApiCodeGen.UI
 {
     [Flags]
     public enum OpenApiTargetFramework