+++ Packages/OpenApiCodeGen/Editor/UI/SettingMenu/View/ISettingView.cs
@@ -1,9 +1,9 @@
 ﻿#nullable enable
 using System;
 
-using ReBeat.OpenApiCodeGen.Core;
+using Rhycol.OpenApiCodeGen.Core;
 
-namespace ReBeat.OpenApiCodeGen.UI
+namespace Rhycol.OpenApiCodeGen.UI
 {
     internal interface ISettingView
     {